Created a model viewer window in Qt, and managed to run the game in it.
I managed to fix the build errors by letting Qt build through QMake, i.e through the Qt Creator interface, and output the binaries to Executables. I then linked ZMasher to the necessary Qt libraries in order to run the window. Making the game run in the Qt window was trivial, I get the HWND from the viewport (QGraphicsView) and use that handle instead of creating a new handle through the windows API. For the widget to run I needed to create a QApplication and run it on a thread parallell to the game. Next task is to link commands between the UI in the modelviewer and the game, probably by using Events. I think I need to create a seperate project in order to create a Modelviewer specific executable, which will include a number of defines for excluding certain runtime code, and enabling events.